feat: Register gRPC server without `@PostConstruct` annotation
resolve #4
