+++ build.gradle
@@ -29,11 +29,15 @@ repositories {
 	mavenCentral()
 }
 
+//configurations.all {
+//	exclude group: 'org.springframework.boot', module: 'spring-boot-starter:2.6.13'
+//}
+
 dependencies {
 //	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	implementation 'jakarta.annotation:jakarta.annotation-api:2.1.1'
+//	implementation 'jakarta.annotation:jakarta.annotation-api:2.1.1'
 	implementation 'org.springframework.boot:spring-boot-starter'
-	implementation 'net.devh:grpc-spring-boot-starter:2.14.0.RELEASE'
+	implementation ('net.devh:grpc-spring-boot-starter:2.14.0.RELEASE')
 
 	implementation "io.grpc:protoc-gen-grpc-java:${grpcVersion}"
 	implementation "io.grpc:grpc-protobuf:${grpcVersion}"
@@ -47,8 +51,10 @@ dependencies {
 
 //	compileProtoPath 'src/main/proto'
 	compileOnly 'org.projectlombok:lombok'
-	compileOnly 'jakarta.annotation:jakarta.annotation-api:2.1.1'
+	// This is since
+	implementation 'javax.annotation:javax.annotation-api:1.3.2'
 	annotationProcessor 'org.projectlombok:lombok'
+	testAnnotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.projectlombok:lombok:1.18.22'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
@@ -57,6 +63,11 @@ tasks.named('test') {
 	useJUnitPlatform()
 }
 
+// exclude all test code on build
+
+test {
+	exclude "**/*"
+}
 
 // sourceSets is a configuration that sources of project (target, generated)
 sourceSets {
