+++ src/main/java/com/example/grpcpractice/server/interceptor/RequestInterceptor.java
@@ -3,11 +3,14 @@ package com.example.grpcpractice.server.interceptor;
 import com.example.grpcpractice.server.metadata.ServerHeaders;
 import io.grpc.*;
 import lombok.extern.slf4j.Slf4j;
+import net.devh.boot.grpc.server.interceptor.GrpcGlobalServerInterceptor;
+import org.springframework.core.annotation.Order;
 
 import java.util.Objects;
 
 @Slf4j
-// This is for validation metadata from client.
+@Order(Integer.MAX_VALUE)
+@GrpcGlobalServerInterceptor
 public class RequestInterceptor implements ServerInterceptor {
     @Override
     public <ReqT, RespT> ServerCall.Listener<ReqT> interceptCall(ServerCall<ReqT, RespT> call, Metadata headers, ServerCallHandler<ReqT, RespT> next) {
@@ -23,17 +26,11 @@ public class RequestInterceptor implements ServerInterceptor {
 
         // ThreadLocal current thread only the information
         // it's safe to use
-//        Context context = Context.current().withValue(
-//                ServerHeaders.CTX_REQUEST_ID,
-//                requestId
-//        );
-
         Context context = Context.current().withValues(
                 ServerHeaders.CTX_REQUEST_ID, requestId,
                 ServerHeaders.CTX_CLIENT_TOKEN, clientToken
         );
 
-//        return next.startCall(call, headers);
         return Contexts.interceptCall(context, call, headers, next);
     }
 }
