+++ src/main/java/com/example/grpcpractice/server/interceptor/ResponseInterceptor.java
@@ -3,9 +3,13 @@ package com.example.grpcpractice.server.interceptor;
 import com.example.grpcpractice.server.metadata.ServerHeaders;
 import io.grpc.*;
 import lombok.extern.slf4j.Slf4j;
+import net.devh.boot.grpc.server.interceptor.GrpcGlobalServerInterceptor;
+import org.springframework.core.annotation.Order;
 
 @Slf4j
+@Order(Integer.MIN_VALUE) // highest priority since first registered intercept is to be executed at last.
 // This is for forwarding metadata after copying data from client.
+@GrpcGlobalServerInterceptor
 public class ResponseInterceptor implements ServerInterceptor {
     @Override
     public <ReqT, RespT> ServerCall.Listener<ReqT> interceptCall(ServerCall<ReqT, RespT>
